refactor: standardizing employee navbar and resolving import issue at adminNavbar
