+++ src/components/AllNavbars/AdminNavbar/AdminNavbar.jsx
@@ -1,13 +1,15 @@
+// IMPORTANTE!!!!!!
+// Essa deve ser a única navbar utilizada p/ todas as telas de Admin
 import { Disclosure, DisclosureButton, DisclosurePanel, Menu, MenuButton, MenuItem, MenuItems } from '@headlessui/react'
 import { Bars3Icon, BellIcon, XMarkIcon } from '@heroicons/react/24/outline'
 import { useState, useEffect } from 'react'
-import getEmployeeData from '../../services/employee/getEmployeeData';
-import profilepic from "../../assets/funcionario/perfil.png";
+import profilepic from "../../../assets/funcionario/perfil.png";
 import { useNavigate } from 'react-router-dom';
+import getAdminData from '../../../services/admin/getAdminData';
 
-const MenuFuncionario = () => {
+const AdminNavbar = () => {
   const navigation = [
-    { name: 'Tickets', href: '/department-tickets', current: true },
+    { name: 'Tickets', href: '/admin-tickets', current: true },
     { name: 'Mensagens', href: '', current: false },
     { name: 'Histórico', href: '', current: false },
   ]
@@ -22,7 +24,7 @@ const MenuFuncionario = () => {
         return;
       }
 
-        let name = await getEmployeeData(sessionStorage.getItem("userId"))
+        let name = await getAdminData(sessionStorage.getItem("userId"))
         if (name) {
             sessionStorage.setItem("username", name)
             setUsername(name);
@@ -65,7 +67,7 @@ const MenuFuncionario = () => {
           </div>
           <div className="flex flex-1 items-center justify-center md:items-stretch md:justify-start">
             <div className="flex flex-shrink-0 items-center ">
-             <h1 onClick={() => navigate("/funcionario")} className='text-white pb-1 text-lg'>WayCliente</h1>
+             <h1 onClick={() => navigate("/adminHome")} className='text-white pb-1 text-lg'>WayClient</h1>
             </div>
             <div className="hidden md:ml-6 md:block ">
               <div className="flex space-x-4 text-white mt-1">
@@ -162,4 +164,4 @@ const MenuFuncionario = () => {
 }
 
 
-export default MenuFuncionario;
+export default AdminNavbar;
