+++ src/pages/AdminTickets.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import MainContent from '../components/AdminTicketsMainContent/MainContent'
-import AdminNavbar from '../components/AdminNavbar/AdminNavbar'
+import AdminNavbar from '../components/AllNavbars/AdminNavbar/AdminNavbar'
 
 const AdminTickets = () => {
   return (
